// backend/src/models/api/StatusResponse.ts
export default interface StatusResponse {
  statusCode: ResponseStatusCode;
  statusMessage: string;
}